requirements.txt
streamlit
playwright
pandas
plotly-express



